+++ src/main/java/com/flab/oasis/service/JwtService.java
@@ -11,17 +11,18 @@ import com.flab.oasis.constant.ErrorCode;
 import com.flab.oasis.constant.JwtProperty;
 import com.flab.oasis.model.JwtToken;
 import com.flab.oasis.model.exception.AuthorizationException;
-import com.flab.oasis.repository.UserAuthRepository;
 import com.flab.oasis.utils.LogUtils;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 
+import java.time.Duration;
 import java.util.Date;
 
 @Service
 @RequiredArgsConstructor
 public class JwtService {
-    private final UserAuthRepository userAuthRepository;
+    private final RedisTemplate<String, Object> redisTemplate;
 
     private static final String ACCESS_TOKEN = "AccessToken";
     private static final String REFRESH_TOKEN = "RefreshToken";
@@ -30,7 +31,11 @@ public class JwtService {
         String accessToken = generateToken(uid, ACCESS_TOKEN);
         String refreshToken = generateToken(uid, REFRESH_TOKEN);
 
-        userAuthRepository.setUserRefreshToken(uid, refreshToken);
+        redisTemplate.opsForValue().set(
+                makeKey(uid),
+                refreshToken,
+                Duration.ofSeconds(JwtProperty.REFRESH_TOKEN_EXPIRE_TIME / 1000)
+        );
 
         return new JwtToken(accessToken, refreshToken);
     }
@@ -38,30 +43,26 @@ public class JwtService {
     public JwtToken reissueJwtToken(String refreshToken) {
         try {
             DecodedJWT decodedJWT = verifyJwt(refreshToken);
+            String uid = decodedJWT.getSubject();
 
             // refresh token이 redis에 존재하지 않으면 다시 로그인을 하도록 한다.
-            if (!userAuthRepository.existUserRefreshToken(decodedJWT.getSubject(), refreshToken)) {
+            if (!refreshToken.equals(String.valueOf(redisTemplate.opsForValue().get(makeKey(uid))))) {
                 throw new AuthorizationException(
                         ErrorCode.NOT_FOUND, "Refresh Token doesn't exist in Redis.", refreshToken
                 );
             }
 
             // refresh token 만료 3일 전이면 토큰을 새로 발급한다.
-            if (willExpire(decodedJWT)) {
-                System.out.println(LogUtils.makeLog(
-                        "Access/Refresh Token wes Reissued.", decodedJWT.getSubject()
-                ));
+            if (willExpire(decodedJWT.getExpiresAt())) {
+                System.out.println(LogUtils.makeLog("Access/Refresh Token wes Reissued.", uid));
 
-                return createJwtToken(decodedJWT.getSubject());
+                return createJwtToken(uid);
             }
 
             // 유효한 refresh token이면 access token만 새로 발급한다.
-            System.out.println(LogUtils.makeLog("Access Token wes Reissued.", decodedJWT.getSubject()));
+            System.out.println(LogUtils.makeLog("Access Token wes Reissued.", uid));
 
-            return new JwtToken(
-                    generateToken(decodedJWT.getSubject(), ACCESS_TOKEN),
-                    refreshToken
-            );
+            return new JwtToken(generateToken(uid, ACCESS_TOKEN), refreshToken);
         } catch (TokenExpiredException e) {
             throw new AuthorizationException(ErrorCode.UNAUTHORIZED, "Refresh Token is Expired.", refreshToken);
         } catch (SignatureVerificationException | InvalidClaimException e) {
@@ -93,7 +94,11 @@ public class JwtService {
                 .sign(algorithm);
     }
 
-    private boolean willExpire(DecodedJWT decodedJWT) {
-        return (decodedJWT.getExpiresAt().getTime() - new Date().getTime()) < JwtProperty.REFRESH_TOKEN_REISSUE_TIME;
+    private String makeKey(String uid) {
+        return String.format("RefreshToken_%s", uid);
+    }
+
+    private boolean willExpire(Date expiresAt) {
+        return (expiresAt.getTime() - new Date().getTime()) < JwtProperty.REFRESH_TOKEN_REISSUE_TIME;
     }
 }
